+++ ostap/fitting/models.py
@@ -130,9 +130,9 @@ __all__ = (
     'Voigt_pdf'            , ## Voigt-profile 
     'PseudoVoigt_pdf'      , ## Voigt-profile 
     'BWI_pdf'              , ## Breit-Wigner with interference 
-    'BW23L_pdf'            , ## BW23L
+    'BW3L_pdf'             , ## BW3L
+    'BWPS_pdf'             , ## BWPS
     'BWMC_pdf'             , ## BWMC
-    'Swanson_pdf'          , ## Swanson's S-wave cusp
     #
     ## "Other" distributions 
     #