+++ ReleaseNotes/release_notes.md
@@ -18,6 +18,7 @@ minuit.minos   ('p1','p2',...)
 minuit.release ('p3')
 ```   
   1. update `pptunnel` + `pplaunch` with better and more informative output 
+  1. add `Ostap::Math::BW3L`, `Ostap::Models::BW3L`,  `BW3L_pdf` and extend  test `ostap/fitting/tests/tests_fitting_breiwigner.py` - resurrected version of the Breit-Wigner profile from 3-body decays 
 
 ## Backward incompatible changes: